Activated the DemographicManager in BoidGraph.
It is actually created with a constant probability of 0.01 for birth and
death for each species.

A bug remains : a species disappear, is this tied to the display and
ui.class ??

+++ src/org/graphstream/boids/BoidGraphListener.java
@@ -44,7 +44,7 @@ public interface BoidGraphListener {
 	 * forces model.
 	 * @param time The current iteration time.
 	 */
-	void step(int time);
+	void step(double time);
 
 	/** 
 	 * One boid was added.
